tests/wpt/web-platform-tests/css/css-flexbox/getcomputedstyle/flexbox_computedstyle_align-items-invalid.html

<!DOCTYPE html>
<title>flexbox | computed style | align-items: invalid</title>
<link rel="author" href="http://opera.com" title="Opera Software">
<link rel="help" href="http://www.w3.org/TR/css-flexbox-1/#align-items-property">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta name="flags" content="dom">
<style>
body {
	color: red;
	align-items: damer;
}
</style>
<div id="log"></div>
<script>
test(function() {
    var body = document.body;

    assert_equals(getComputedStyle(body).getPropertyValue("align-items"),
	          "normal");
});
</script>
